Make board reconnect to MQTT broker if connection is lost.

+++ src/VentilationFan.h
@@ -8,13 +8,14 @@
 #ifndef VENTILATIONFAN_H_
 #define VENTILATIONFAN_H_
 
+#include "ModbusRegister.h"
+
 class VentilationFan {
 public:
-	VentilationFan(ModbusRegister *readRegister, ModbusRegister *writeRegister,
-			int speed, bool spinning);
+	VentilationFan(ModbusRegister *readRegister, ModbusRegister *writeRegister,int speed, bool spinning);
 	virtual ~VentilationFan();
 	bool readFan();
-	void setSpeed(int speed);
+	void setSpeed(float speed);
 private:
 	int speed;
 	bool spinning;
